+++ api/types/storage/driver_data.go
@@ -1,13 +1,13 @@
-package types
+package storage
 
 // This file was generated by the swagger tool.
 // Editing this file might prove futile when you re-run the swagger generate command
 
-// GraphDriverData Information about the storage driver used to store the container's and
+// DriverData Information about the storage driver used to store the container's and
 // image's filesystem.
 //
-// swagger:model GraphDriverData
-type GraphDriverData struct {
+// swagger:model DriverData
+type DriverData struct {
 
 	// Low-level storage metadata, provided as key/value pairs.
 	//
