+++ hack/generate-swagger-api.sh
@@ -4,13 +4,16 @@ set -eu
 swagger generate model -f api/swagger.yaml \
 	-t api -m types --skip-validator -C api/swagger-gen.yaml \
 	-n ErrorResponse \
-	-n GraphDriverData \
 	-n IdResponse \
 	-n Plugin \
 	-n PluginDevice \
 	-n PluginMount \
 	-n PluginEnv \
 	-n PluginInterfaceType
+
+swagger generate model -f api/swagger.yaml \
+	-t api -m types/storage --skip-validator -C api/swagger-gen.yaml \
+	-n DriverData
 
 swagger generate model -f api/swagger.yaml \
 	-t api -m types/container --skip-validator -C api/swagger-gen.yaml \
